Print what install.sh is doing
Also removed a few old dotfile refs that aren't in the repo any more,
and which the new output highlighted.

+++ install.sh
@@ -18,8 +18,6 @@ readonly DOTDIRECTORIES="
 "
 
 readonly DOTFILES="
-  ackrc
-  bash_aliases
   bash_completion
   bashrc
   env_hooker
@@ -27,7 +25,6 @@ readonly DOTFILES="
   gitattributes
   gitconfig
   gitignore_global
-  gvimrc
   inputrc
   irbrc
   npmrc
@@ -39,25 +36,42 @@ readonly DOTFILES="
 "
 
 for dir in $DOTDIRECTORIES; do
-  if [[ -n "$dir" ]]; then
+  if [[ ! -d "$dir" ]]; then
+    echo "Symlinking: ${HOME}/dotfiles/${dir} => ${HOME}/.${dir}"
     $dircmd "${HOME}/dotfiles/${dir}" "${HOME}/.${dir}"
+  else
+    echo "    Exists: ${HOME}/.${dir}"
   fi
 done
 
 for file in $DOTFILES; do
-  if [[ -n "$file" ]]; then
+  if [[ ! -f "$file" ]]; then
+    echo "Symlinking: ${HOME}/dotfiles/${file} => ${HOME}/.${file}"
     ln -sf "${HOME}/dotfiles/${file}" "${HOME}/.${file}"
+  else
+
+    echo "    Exists: ${HOME}/.${file}"
   fi
 done
 
 mkdir -p ~/.bashrc.d
 for script_file in ~/dotfiles/bashrc.d/*; do
   script_name=$(basename "${script_file}")
-  ln -sf "${script_file}" "${HOME}/.bashrc.d/${script_name}"
+  if [[ ! -f "${script_file}" ]]; then
+    echo "Symlinking: ${script_file} => ${HOME}/.bashrc.d/${script_name}"
+    ln -sf "${script_file}" "${HOME}/.bashrc.d/${script_name}"
+  else
+    echo "    Exists: ${HOME}/.bashrc.d/${script_name}"
+  fi
 done
 
 mkdir -p ~/bin
 for script_file in ~/dotfiles/bin/*; do
   script_name=$(basename "${script_file}")
-  ln -sf "${script_file}" "${HOME}/bin/${script_name}"
+  if [[ ! -f "${script_file}" ]]; then
+    echo "Symlinking: ${script_file} => ${HOME}/bin/${script_name}"
+    ln -sf "${script_file}" "${HOME}/bin/${script_name}"
+  else
+    echo "    Exists: ${HOME}/bin/${script_name}"
+  fi
 done
